+++ lib/geo_triggering_page.dart
@@ -110,9 +110,29 @@ class _GeoTriggeringPageState extends State<GeoTriggeringPage> {
     });
   }
 
+  /// Set Notification Icon (for Android)
+  void _setNotificationIcon() {
+    // set to the notification in your drawable folder, without the extension
+    BluedotPointSdk.instance.setNotificationIcon('ic_stat_name').then((value) {
+      // Successfully stop geo triggering
+      _updateGeoTriggeringStatus();
+    }).catchError((error) {
+      // Failed to stop geo triggering, handle error in here
+      String errorMessage = error.toString();
+      if (error is PlatformException) {
+        errorMessage = error.message!;
+      }
+      showAlert('Fail to set notification icon', errorMessage, context);
+    });
+  }
+
   @override
   void initState() {
     super.initState();
+
+    /// Make sure Notification Icon is set prior to starting GeoTriggering
+    /// if you want to set Foreground Notification icon for Android
+    _setNotificationIcon();
 
     // Handle geo triggering events
     geoTriggeringEventChannel.setMethodCallHandler((MethodCall call) async {